Split PersonalizedTotals in other file
